Fix multiple definitions
This commit fixes multiple definitions issues in the caffe2 test harness.
This was caught internally when building on FB Infra.
